requirements.txt
tensorflow==2.15.0
matplotlib
numpy
scikit-learn
Pillow